init promql optimizer
