+++ promql/engine_test.go
@@ -1587,724 +1587,6 @@ func TestQueryLogger_error(t *testing.T) {
 	}
 }
 
-func TestPreprocessAndWrapWithStepInvariantExpr(t *testing.T) {
-	startTime := time.Unix(1000, 0)
-	endTime := time.Unix(9999, 0)
-	testCases := []struct {
-		input    string      // The input to be parsed.
-		expected parser.Expr // The expected expression AST.
-	}{
-		{
-			input: "123.4567",
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.NumberLiteral{
-					Val:      123.4567,
-					PosRange: parser.PositionRange{Start: 0, End: 8},
-				},
-			},
-		}, {
-			input: `"foo"`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.StringLiteral{
-					Val:      "foo",
-					PosRange: parser.PositionRange{Start: 0, End: 5},
-				},
-			},
-		}, {
-			input: "foo * bar",
-			expected: &parser.BinaryExpr{
-				Op: parser.MUL,
-				LHS: &parser.VectorSelector{
-					Name: "foo",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   3,
-					},
-				},
-				RHS: &parser.VectorSelector{
-					Name: "bar",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "bar"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 6,
-						End:   9,
-					},
-				},
-				VectorMatching: &parser.VectorMatching{Card: parser.CardOneToOne},
-			},
-		}, {
-			input: "foo * bar @ 10",
-			expected: &parser.BinaryExpr{
-				Op: parser.MUL,
-				LHS: &parser.VectorSelector{
-					Name: "foo",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   3,
-					},
-				},
-				RHS: &parser.StepInvariantExpr{
-					Expr: &parser.VectorSelector{
-						Name: "bar",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "bar"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 6,
-							End:   14,
-						},
-						Timestamp: makeInt64Pointer(10000),
-					},
-				},
-				VectorMatching: &parser.VectorMatching{Card: parser.CardOneToOne},
-			},
-		}, {
-			input: "foo @ 20 * bar @ 10",
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.BinaryExpr{
-					Op: parser.MUL,
-					LHS: &parser.VectorSelector{
-						Name: "foo",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   8,
-						},
-						Timestamp: makeInt64Pointer(20000),
-					},
-					RHS: &parser.VectorSelector{
-						Name: "bar",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "bar"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 11,
-							End:   19,
-						},
-						Timestamp: makeInt64Pointer(10000),
-					},
-					VectorMatching: &parser.VectorMatching{Card: parser.CardOneToOne},
-				},
-			},
-		}, {
-			input: "test[5s]",
-			expected: &parser.MatrixSelector{
-				VectorSelector: &parser.VectorSelector{
-					Name: "test",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "test"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   4,
-					},
-				},
-				Range:  5 * time.Second,
-				EndPos: 8,
-			},
-		}, {
-			input: `test{a="b"}[5y] @ 1603774699`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.MatrixSelector{
-					VectorSelector: &parser.VectorSelector{
-						Name:      "test",
-						Timestamp: makeInt64Pointer(1603774699000),
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "a", "b"),
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "test"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   11,
-						},
-					},
-					Range:  5 * 365 * 24 * time.Hour,
-					EndPos: 28,
-				},
-			},
-		}, {
-			input: "sum by (foo)(some_metric)",
-			expected: &parser.AggregateExpr{
-				Op: parser.SUM,
-				Expr: &parser.VectorSelector{
-					Name: "some_metric",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 13,
-						End:   24,
-					},
-				},
-				Grouping: []string{"foo"},
-				PosRange: parser.PositionRange{
-					Start: 0,
-					End:   25,
-				},
-			},
-		}, {
-			input: "sum by (foo)(some_metric @ 10)",
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.AggregateExpr{
-					Op: parser.SUM,
-					Expr: &parser.VectorSelector{
-						Name: "some_metric",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 13,
-							End:   29,
-						},
-						Timestamp: makeInt64Pointer(10000),
-					},
-					Grouping: []string{"foo"},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   30,
-					},
-				},
-			},
-		}, {
-			input: "sum(some_metric1 @ 10) + sum(some_metric2 @ 20)",
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.BinaryExpr{
-					Op:             parser.ADD,
-					VectorMatching: &parser.VectorMatching{},
-					LHS: &parser.AggregateExpr{
-						Op: parser.SUM,
-						Expr: &parser.VectorSelector{
-							Name: "some_metric1",
-							LabelMatchers: []*labels.Matcher{
-								parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric1"),
-							},
-							PosRange: parser.PositionRange{
-								Start: 4,
-								End:   21,
-							},
-							Timestamp: makeInt64Pointer(10000),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   22,
-						},
-					},
-					RHS: &parser.AggregateExpr{
-						Op: parser.SUM,
-						Expr: &parser.VectorSelector{
-							Name: "some_metric2",
-							LabelMatchers: []*labels.Matcher{
-								parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric2"),
-							},
-							PosRange: parser.PositionRange{
-								Start: 29,
-								End:   46,
-							},
-							Timestamp: makeInt64Pointer(20000),
-						},
-						PosRange: parser.PositionRange{
-							Start: 25,
-							End:   47,
-						},
-					},
-				},
-			},
-		}, {
-			input: "some_metric and topk(5, rate(some_metric[1m] @ 20))",
-			expected: &parser.BinaryExpr{
-				Op: parser.LAND,
-				VectorMatching: &parser.VectorMatching{
-					Card: parser.CardManyToMany,
-				},
-				LHS: &parser.VectorSelector{
-					Name: "some_metric",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   11,
-					},
-				},
-				RHS: &parser.StepInvariantExpr{
-					Expr: &parser.AggregateExpr{
-						Op: parser.TOPK,
-						Expr: &parser.Call{
-							Func: parser.MustGetFunction("rate"),
-							Args: parser.Expressions{
-								&parser.MatrixSelector{
-									VectorSelector: &parser.VectorSelector{
-										Name: "some_metric",
-										LabelMatchers: []*labels.Matcher{
-											parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric"),
-										},
-										PosRange: parser.PositionRange{
-											Start: 29,
-											End:   40,
-										},
-										Timestamp: makeInt64Pointer(20000),
-									},
-									Range:  1 * time.Minute,
-									EndPos: 49,
-								},
-							},
-							PosRange: parser.PositionRange{
-								Start: 24,
-								End:   50,
-							},
-						},
-						Param: &parser.NumberLiteral{
-							Val: 5,
-							PosRange: parser.PositionRange{
-								Start: 21,
-								End:   22,
-							},
-						},
-						PosRange: parser.PositionRange{
-							Start: 16,
-							End:   51,
-						},
-					},
-				},
-			},
-		}, {
-			input: "time()",
-			expected: &parser.Call{
-				Func: parser.MustGetFunction("time"),
-				Args: parser.Expressions{},
-				PosRange: parser.PositionRange{
-					Start: 0,
-					End:   6,
-				},
-			},
-		}, {
-			input: `foo{bar="baz"}[10m:6s]`,
-			expected: &parser.SubqueryExpr{
-				Expr: &parser.VectorSelector{
-					Name: "foo",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "bar", "baz"),
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   14,
-					},
-				},
-				Range:  10 * time.Minute,
-				Step:   6 * time.Second,
-				EndPos: 22,
-			},
-		}, {
-			input: `foo{bar="baz"}[10m:6s] @ 10`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.SubqueryExpr{
-					Expr: &parser.VectorSelector{
-						Name: "foo",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "bar", "baz"),
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   14,
-						},
-					},
-					Range:     10 * time.Minute,
-					Step:      6 * time.Second,
-					Timestamp: makeInt64Pointer(10000),
-					EndPos:    27,
-				},
-			},
-		}, { // Even though the subquery is step invariant, the inside is also wrapped separately.
-			input: `sum(foo{bar="baz"} @ 20)[10m:6s] @ 10`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.SubqueryExpr{
-					Expr: &parser.StepInvariantExpr{
-						Expr: &parser.AggregateExpr{
-							Op: parser.SUM,
-							Expr: &parser.VectorSelector{
-								Name: "foo",
-								LabelMatchers: []*labels.Matcher{
-									parser.MustLabelMatcher(labels.MatchEqual, "bar", "baz"),
-									parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-								},
-								PosRange: parser.PositionRange{
-									Start: 4,
-									End:   23,
-								},
-								Timestamp: makeInt64Pointer(20000),
-							},
-							PosRange: parser.PositionRange{
-								Start: 0,
-								End:   24,
-							},
-						},
-					},
-					Range:     10 * time.Minute,
-					Step:      6 * time.Second,
-					Timestamp: makeInt64Pointer(10000),
-					EndPos:    37,
-				},
-			},
-		}, {
-			input: `min_over_time(rate(foo{bar="baz"}[2s])[5m:] @ 1603775091)[4m:3s]`,
-			expected: &parser.SubqueryExpr{
-				Expr: &parser.StepInvariantExpr{
-					Expr: &parser.Call{
-						Func: parser.MustGetFunction("min_over_time"),
-						Args: parser.Expressions{
-							&parser.SubqueryExpr{
-								Expr: &parser.Call{
-									Func: parser.MustGetFunction("rate"),
-									Args: parser.Expressions{
-										&parser.MatrixSelector{
-											VectorSelector: &parser.VectorSelector{
-												Name: "foo",
-												LabelMatchers: []*labels.Matcher{
-													parser.MustLabelMatcher(labels.MatchEqual, "bar", "baz"),
-													parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-												},
-												PosRange: parser.PositionRange{
-													Start: 19,
-													End:   33,
-												},
-											},
-											Range:  2 * time.Second,
-											EndPos: 37,
-										},
-									},
-									PosRange: parser.PositionRange{
-										Start: 14,
-										End:   38,
-									},
-								},
-								Range:     5 * time.Minute,
-								Timestamp: makeInt64Pointer(1603775091000),
-								EndPos:    56,
-							},
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   57,
-						},
-					},
-				},
-				Range:  4 * time.Minute,
-				Step:   3 * time.Second,
-				EndPos: 64,
-			},
-		}, {
-			input: `some_metric @ 123 offset 1m [10m:5s]`,
-			expected: &parser.SubqueryExpr{
-				Expr: &parser.StepInvariantExpr{
-					Expr: &parser.VectorSelector{
-						Name: "some_metric",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   27,
-						},
-						Timestamp:      makeInt64Pointer(123000),
-						OriginalOffset: 1 * time.Minute,
-					},
-				},
-				Range:  10 * time.Minute,
-				Step:   5 * time.Second,
-				EndPos: 36,
-			},
-		}, {
-			input: `some_metric[10m:5s] offset 1m @ 123`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.SubqueryExpr{
-					Expr: &parser.VectorSelector{
-						Name: "some_metric",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   11,
-						},
-					},
-					Timestamp:      makeInt64Pointer(123000),
-					OriginalOffset: 1 * time.Minute,
-					Range:          10 * time.Minute,
-					Step:           5 * time.Second,
-					EndPos:         35,
-				},
-			},
-		}, {
-			input: `(foo + bar{nm="val"} @ 1234)[5m:] @ 1603775019`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.SubqueryExpr{
-					Expr: &parser.ParenExpr{
-						Expr: &parser.BinaryExpr{
-							Op: parser.ADD,
-							VectorMatching: &parser.VectorMatching{
-								Card: parser.CardOneToOne,
-							},
-							LHS: &parser.VectorSelector{
-								Name: "foo",
-								LabelMatchers: []*labels.Matcher{
-									parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-								},
-								PosRange: parser.PositionRange{
-									Start: 1,
-									End:   4,
-								},
-							},
-							RHS: &parser.StepInvariantExpr{
-								Expr: &parser.VectorSelector{
-									Name: "bar",
-									LabelMatchers: []*labels.Matcher{
-										parser.MustLabelMatcher(labels.MatchEqual, "nm", "val"),
-										parser.MustLabelMatcher(labels.MatchEqual, "__name__", "bar"),
-									},
-									Timestamp: makeInt64Pointer(1234000),
-									PosRange: parser.PositionRange{
-										Start: 7,
-										End:   27,
-									},
-								},
-							},
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   28,
-						},
-					},
-					Range:     5 * time.Minute,
-					Timestamp: makeInt64Pointer(1603775019000),
-					EndPos:    46,
-				},
-			},
-		}, {
-			input: "abs(abs(metric @ 10))",
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.Call{
-					Func: &parser.Function{
-						Name:       "abs",
-						ArgTypes:   []parser.ValueType{parser.ValueTypeVector},
-						ReturnType: parser.ValueTypeVector,
-					},
-					Args: parser.Expressions{&parser.Call{
-						Func: &parser.Function{
-							Name:       "abs",
-							ArgTypes:   []parser.ValueType{parser.ValueTypeVector},
-							ReturnType: parser.ValueTypeVector,
-						},
-						Args: parser.Expressions{&parser.VectorSelector{
-							Name: "metric",
-							LabelMatchers: []*labels.Matcher{
-								parser.MustLabelMatcher(labels.MatchEqual, "__name__", "metric"),
-							},
-							PosRange: parser.PositionRange{
-								Start: 8,
-								End:   19,
-							},
-							Timestamp: makeInt64Pointer(10000),
-						}},
-						PosRange: parser.PositionRange{
-							Start: 4,
-							End:   20,
-						},
-					}},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   21,
-					},
-				},
-			},
-		}, {
-			input: "sum(sum(some_metric1 @ 10) + sum(some_metric2 @ 20))",
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.AggregateExpr{
-					Op: parser.SUM,
-					Expr: &parser.BinaryExpr{
-						Op:             parser.ADD,
-						VectorMatching: &parser.VectorMatching{},
-						LHS: &parser.AggregateExpr{
-							Op: parser.SUM,
-							Expr: &parser.VectorSelector{
-								Name: "some_metric1",
-								LabelMatchers: []*labels.Matcher{
-									parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric1"),
-								},
-								PosRange: parser.PositionRange{
-									Start: 8,
-									End:   25,
-								},
-								Timestamp: makeInt64Pointer(10000),
-							},
-							PosRange: parser.PositionRange{
-								Start: 4,
-								End:   26,
-							},
-						},
-						RHS: &parser.AggregateExpr{
-							Op: parser.SUM,
-							Expr: &parser.VectorSelector{
-								Name: "some_metric2",
-								LabelMatchers: []*labels.Matcher{
-									parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric2"),
-								},
-								PosRange: parser.PositionRange{
-									Start: 33,
-									End:   50,
-								},
-								Timestamp: makeInt64Pointer(20000),
-							},
-							PosRange: parser.PositionRange{
-								Start: 29,
-								End:   52,
-							},
-						},
-					},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   52,
-					},
-				},
-			},
-		}, {
-			input: `foo @ start()`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.VectorSelector{
-					Name: "foo",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   13,
-					},
-					Timestamp:  makeInt64Pointer(timestamp.FromTime(startTime)),
-					StartOrEnd: parser.START,
-				},
-			},
-		}, {
-			input: `foo @ end()`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.VectorSelector{
-					Name: "foo",
-					LabelMatchers: []*labels.Matcher{
-						parser.MustLabelMatcher(labels.MatchEqual, "__name__", "foo"),
-					},
-					PosRange: parser.PositionRange{
-						Start: 0,
-						End:   11,
-					},
-					Timestamp:  makeInt64Pointer(timestamp.FromTime(endTime)),
-					StartOrEnd: parser.END,
-				},
-			},
-		}, {
-			input: `test[5y] @ start()`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.MatrixSelector{
-					VectorSelector: &parser.VectorSelector{
-						Name:       "test",
-						Timestamp:  makeInt64Pointer(timestamp.FromTime(startTime)),
-						StartOrEnd: parser.START,
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "test"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   4,
-						},
-					},
-					Range:  5 * 365 * 24 * time.Hour,
-					EndPos: 18,
-				},
-			},
-		}, {
-			input: `test[5y] @ end()`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.MatrixSelector{
-					VectorSelector: &parser.VectorSelector{
-						Name:       "test",
-						Timestamp:  makeInt64Pointer(timestamp.FromTime(endTime)),
-						StartOrEnd: parser.END,
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "test"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   4,
-						},
-					},
-					Range:  5 * 365 * 24 * time.Hour,
-					EndPos: 16,
-				},
-			},
-		}, {
-			input: `some_metric[10m:5s] @ start()`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.SubqueryExpr{
-					Expr: &parser.VectorSelector{
-						Name: "some_metric",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   11,
-						},
-					},
-					Timestamp:  makeInt64Pointer(timestamp.FromTime(startTime)),
-					StartOrEnd: parser.START,
-					Range:      10 * time.Minute,
-					Step:       5 * time.Second,
-					EndPos:     29,
-				},
-			},
-		}, {
-			input: `some_metric[10m:5s] @ end()`,
-			expected: &parser.StepInvariantExpr{
-				Expr: &parser.SubqueryExpr{
-					Expr: &parser.VectorSelector{
-						Name: "some_metric",
-						LabelMatchers: []*labels.Matcher{
-							parser.MustLabelMatcher(labels.MatchEqual, "__name__", "some_metric"),
-						},
-						PosRange: parser.PositionRange{
-							Start: 0,
-							End:   11,
-						},
-					},
-					Timestamp:  makeInt64Pointer(timestamp.FromTime(endTime)),
-					StartOrEnd: parser.END,
-					Range:      10 * time.Minute,
-					Step:       5 * time.Second,
-					EndPos:     27,
-				},
-			},
-		},
-	}
-
-	for _, test := range testCases {
-		t.Run(test.input, func(t *testing.T) {
-			expr, err := parser.ParseExpr(test.input)
-			require.NoError(t, err)
-			expr = PreprocessExpr(expr, startTime, endTime)
-			require.Equal(t, test.expected, expr, "error on input '%s'", test.input)
-		})
-	}
-}
-
 func TestEngineOptsValidation(t *testing.T) {
 	cases := []struct {
 		opts     EngineOpts
